+++ test-files/test-2.ts
@@ -1,0 +1,4 @@
+import { TypeScriptTest } from './test';
+
+const anotherInstance = new TypeScriptTest();
+console.log("Another one: " + anotherInstance.hello());